% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{lda}
\alias{lda}
\title{Linear Discriminant Analysis using R}
\usage{
lda(X, Y, d)
}
\arguments{
\item{X}{the Sample matrix(numeric)}

\item{Y}{the Sample matrix(numeric)}

\item{d}{the number of dimension of sample matrix}
}
\value{
a random sample list
}
\description{
Use R package \code{lda} to achieve data dimensionality reduction
}
