% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{lle}
\alias{lle}
\title{Locally linear embedding using R}
\usage{
lle(X, d, k)
}
\arguments{
\item{X}{the Sample matrix(numeric)}

\item{d}{the number of dimension of sample matrix}

\item{k}{the number of nearest neighbors}
}
\value{
a random sample matrix
}
\description{
Use R package \code{lle} to achieve data dimensionality reduction
}
