% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{pca}
\alias{pca}
\title{Principal component analysis using R}
\usage{
pca(X)
}
\arguments{
\item{X}{the Sample matrix(numeric)}
}
\value{
a random sample list
}
\description{
Use R package \code{pca} to achieve data dimensionality reduction of sample matrix
}
