% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{Metropolis samples using Rcpp}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the sample variance}

\item{x0}{the given initial value}

\item{N}{the number of the samples}
}
\value{
a random sample of size \code{n}
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
sigma = 2
x0 = 25
N = 2000
rwMetropolis(sigma,x0,N)
}
}
