\name{StatComp20039-package}
\alias{StatComp20039-package}
\docType{package}
\title{
  This package provides several functions for dimensionality reduction of matrix data.
}
\description{
  R functions of dimensionality reduction of matrix data are provided in this package (\code{pca} and \code{lda} and \code{lle}).For high-dimensional data, the dimension reduction method is used to reduce the feature dimension to 2-D or 3-D.
}
\author{
xinyi C

Maintainer: xinyi Cao<xinyicao@mail.ustc.edu.cn>
}
\keyword{ package }
